requirements.txt
pyaudio
speechrecognition
groq
python-dotenv

